timeout a geolocation request
it seems in airplane mode the error callback is never called
so we should set a timeout so this gives up looking after some time
(20s for the time being - random choice of number)

note allowing the user to 'go back' is problematic as the request could
still succeed and the callback might fire after the cancel. Pretty sure you
cannot cancel a getCurrentPosition request but may be wrong.